+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # PayPal iOS SDK Release Notes
 
+## unreleased
+* PaymentButtons
+  * Add black boundary around white buttons
+
 ## 1.3.1 (2024-04-29)
 * FraudProtection
   * Include `DeviceID` privacy term in `PrivacyInfo.xcprivacy` file